refactor(es): Simplify summary ordering and split ReaderWithSummaries into summary.go
